Fix 7-puts_half.c

+++ pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
 void puts_half(char *str)
 {
 	int i;
-	int length = _strlen(str);
+	int length = (int)_strlen(str);
 
 	for (i = length / 2; i < length; i++)
 	{
